Simple authentication and patient registration (#19)
Created a patient controller endpoint in the backend that allows registration of a user and the creation of a patient document in the database

Creation of a patient domain model that stores the device tokens associated with that patient.

Login and session management are handled using the firebase sdk in the frontend. Very minimal error handling and validation, but it works.

Very simple login and create account ui added (still needs some work though)

spring-api.ts in the frontend is where we will make all our calls to the backend